requirements.txt
# ─── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6

# ─── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3

# ─── Data Validation ───────────────────────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# ─── Authentication ────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12         # Required for OAuth2PasswordRequestForm

# ─── Email ─────────────────────────────────────────────────────────────────────
# Uses Python built-in smtplib — no extra package needed

# ─── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1

# ─── Development / Testing ─────────────────────────────────────────────────────
pytest==8.3.3
pytest-cov==5.0.0
httpx==0.27.2                    # For FastAPI TestClient
